feat: カメラと音声のレコーディング処理を、Activity から分離し専用クラスに
